feat: Update project name and enhance default cities
Updates the project name from "aero-stat" to "aerostatv2" in package.json and metadata.json. Also, expands the default list of cities to include more international locations and refactors the OpenWeather API key to use a constant. The title in index.html is also updated for better branding.

Removed CNAME file as it was empty.